# specs/api.openapi.yml
openapi: 3.0.0
info:
  title: Trading Platform API - iQOption Integration
  description: REST API for iQOption broker connection, authentication, and account management
  version: 1.0.0
  contact:
    name: Trading Platform Team
  license:
    name: MIT

servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.trading-platform.com
    description: Production server

paths:
  /api/broker/login:
    post:
      operationId: brokerLogin
      summary: Authenticate with iQOption broker
      description: Authenticates user with iQOption broker using credentials and returns session token
      tags:
        - Broker
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: iQOption account email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  description: iQOption account password
              required:
                - email
                - password
      responses:
        '200':
          description: Successfully authenticated with iQOption
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  sessionId:
                    type: string
                    description: Session token for authenticated requests
                  broker:
                    type: string
                    example: iQOption
                  mode:
                    type: string
                    enum: [demo, live]
                    description: Current trading mode
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Authentication failed - invalid credentials
        '400':
          description: Missing required fields (email or password)

  /api/broker/logout:
    post:
      operationId: brokerLogout
      summary: Logout from iQOption broker
      description: Terminates session and cleans up broker connection
      tags:
        - Broker
        - Authentication
      responses:
        '200':
          description: Successfully logged out from iQOption
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Not authenticated or session expired

  /api/broker/balance:
    get:
      operationId: getBalance
      summary: Get current account balance
      description: Retrieves account balance from iQOption for current mode (demo/live)
      tags:
        - Broker
        - Account
      responses:
        '200':
          description: Successfully retrieved balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  balance:
                    type: number
                    format: float
                    description: Current account balance
                  currency:
                    type: string
                    description: Currency code
                    example: USD
                  mode:
                    type: string
                    enum: [demo, live]
                    description: Current trading mode
                  lastUpdated:
                    type: string
                    format: date-time
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Not authenticated - no active session
        '500':
          description: Error retrieving balance from broker

  /api/broker/mode/switch:
    post:
      operationId: switchMode
      summary: Switch trading mode
      description: Switches between demo and live trading modes
      tags:
        - Broker
        - Account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mode:
                  type: string
                  enum: [demo, live]
                  description: Target trading mode
              required:
                - mode
      responses:
        '200':
          description: Successfully switched trading mode
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  mode:
                    type: string
                    enum: [demo, live]
                    description: Current trading mode
                  balance:
                    type: number
                    format: float
                    description: Balance for the new mode
                  currency:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Invalid mode specified (not demo or live)
        '401':
          description: Not authenticated - no active session
        '500':
          description: Error switching mode with broker

components:
  schemas:
    BrokerSession:
      type: object
      properties:
        sessionId:
          type: string
        broker:
          type: string
        email:
          type: string
        mode:
          type: string
          enum: [demo, live]
        timestamp:
          type: string
          format: date-time

    Balance:
      type: object
      properties:
        balance:
          type: number
        currency:
          type: string
        mode:
          type: string
          enum: [demo, live]
        lastUpdated:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string
        timestamp:
          type: string
          format: date-time